License added and enum.py file created instead utils
